+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- properties to support SharePoint Create List Item form workflow event
+
 ## [10.1.1] - 2025-01-22
 
 ### Added